port2: Replace short with int for non-disk-format variables (plan item 8)
The short type was a memory-saving measure for old machines. Replace
with int for local variables, parameters, and globals that are not
part of on-disk structures. Struct fields in ex_temp.c that define
the register file disk format are intentionally left as short.

+++ port2/ex.h
@@ -158,9 +158,9 @@ extern	int	xchng;		/* Suppresses multiple "No writes" in !cmd */
 extern	char	termtype[];	/* Terminal type from strstrstrstrstrstring capability */
 extern	bool	pflag, nflag;	/* Print/number flags for commands */
 extern	int	poffset;	/* Print line number offset */
-extern	short	isalt;		/* Current file is alternate */
-extern	short	slevel;		/* Source nesting level */
-extern	short	ttyindes;	/* TTY input descriptor */
+extern	int	isalt;		/* Current file is alternate */
+extern	int	slevel;		/* Source nesting level */
+extern	int	ttyindes;	/* TTY input descriptor */
 extern	int	verbose;	/* Verbose option (ex -V) */
 extern	void	oncore(int);	/* Core dump signal handler */
